fix: Workflow transition sequence.
